use test_client for all server testing
